+++ autosbatch/autosbatch.py
@@ -1,49 +1,47 @@
 """Main module."""
+import datetime
+import json
 import logging
 import os
 import time
 from collections import OrderedDict
 from pathlib import Path
-from subprocess import PIPE, call, check_output, run
+from subprocess import PIPE, run
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
 from jinja2 import Environment, FileSystemLoader
-from rich.logging import RichHandler
 
-FORMAT = "%(message)s"
-logging.basicConfig(level="NOTSET", format=FORMAT, datefmt="[%X]", handlers=[RichHandler()])
+from .logger import logger
 
-logger = logging.getLogger(__name__)
-
-# TODO: Add a logger
 # TODO: Add a progress bar
-# TODO: submit to nodes sorted by load
 
 
 class SlurmPool:
     """A class for submitting jobs to Slurm."""
 
     # TODO: support context manager
-    FILE_DIR = './.autosbatch'
+    time_now = datetime.datetime.now().strftime('%m%d%H%M%S')
+    FILE_DIR = f'./.autosbatch/{time_now}'
     LOG_DIR = f'{FILE_DIR}/log'
     SCRIPTS_DIR = f'{FILE_DIR}/scripts'
+
+    CPU_OpenMP_TEMPLATE = 'CPU_OpenMP.j2'
 
     def __init__(
         self,
         pool_size: Optional[int] = None,
-        ncpus_per_job: int = 2,
+        ncpus_per_job: int = 1,
         max_jobs_per_node: Optional[int] = None,
         node_list: Optional[List[str]] = None,
         partition: Optional[str] = None,
         max_pool_size: int = 1000,
     ):
-        self.ncpus_per_job = ncpus_per_job  # TODO: change to 1, when hyperthreading is disabled
+        self.ncpus_per_job = ncpus_per_job
 
         self.nodes = self.get_nodes()
         self._get_avail_nodes(node_list=node_list, partition=partition)
         self.node_list = list(self.nodes.keys())
         if len(self.node_list) == 0:
-            # TODO: use logging.error
             raise RuntimeError('No Nodes are qualtified.')
 
         self._set_max_jobs_per_node(max_jobs_per_node=max_jobs_per_node)
@@ -57,7 +55,6 @@ class SlurmPool:
     @classmethod
     def get_nodes(cls, sortByload=True) -> OrderedDict:
         """Get nodes information from sinfo."""
-        # TODO: get rid of pandas
         command = ['sinfo', '-o', '"%n %e %m %a %c %C %O %R %t"']
         result = run(command, stdout=PIPE, stderr=PIPE, universal_newlines=True)
         nodes = {}
@@ -82,8 +79,16 @@ class SlurmPool:
             nodes = OrderedDict(
                 sorted(nodes.items(), key=lambda x: (x[1]['load'], x[1]['used_mem'], x[1]['used_cpus']))
             )
-        logger.debug(f'Found {len(nodes)} nodes.')
         return nodes
+
+    def _check_hypertreading(self, node_name) -> bool:
+        """Check if hyperthreading is enabled."""
+        command = ['scontrol', 'show', 'node', node_name]
+        result = run(command, stdout=PIPE, stderr=PIPE, universal_newlines=True)
+        if 'ThreadsPerCore=2' in result.stdout:
+            return True
+        else:
+            return False
 
     def _get_avail_nodes(
         self,
@@ -103,6 +108,14 @@ class SlurmPool:
         self, max_jobs_per_node: Optional[int] = None,
     ):
         """Set max_jobs_per_node."""
+        if self.ncpus_per_job & 0x1:
+            for node in self.nodes.keys():
+                if self._check_hypertreading(node):
+                    self.ncpus_per_job += 1
+                    logger.warning(
+                        f'Hyperthreading is enabled on {node}, ncpus_per_job is set to {self.ncpus_per_job}.'
+                    )
+                    break
         for v in self.nodes.values():
             v['max_jobs'] = v['free_cpus'] // self.ncpus_per_job
         avail_max_jobs_per_node = max(v['max_jobs'] for v in self.nodes.values())
@@ -135,55 +148,94 @@ class SlurmPool:
         cpus_per_task: int,
         cmds: List[str],
         job_name: str = 'aotusbatch',
-        job_id: str = '001',
         logging_level: int = logging.WARNING,
     ):
-        # TODO: replace call with subprocess.run
+        logger.setLevel(logging_level)
         Path(cls.SCRIPTS_DIR).mkdir(parents=True, exist_ok=True)
         Path(cls.LOG_DIR).mkdir(parents=True, exist_ok=True)
         templateLoader = FileSystemLoader(searchpath=f"{os.path.dirname(os.path.realpath(__file__))}/template")
         env = Environment(loader=templateLoader)
-        template = env.get_template('CPU_OpenMP.j2')
+        template = env.get_template(cls.CPU_OpenMP_TEMPLATE)
         output_from_parsed_template = template.render(
             job_name=job_name,
-            job_id=job_id,
             partition=partition,
             node=node,
             cpus_per_task=cpus_per_task,
             cmds='\n'.join(cmds),
+            log_dir=cls.LOG_DIR,
         )
-        with open(f'{cls.SCRIPTS_DIR}/scripts_{job_id}.sh', 'w') as f:
+        script_path = f'{cls.SCRIPTS_DIR}/{job_name}.sh'
+        with open(script_path, 'w') as f:
             f.write(output_from_parsed_template)
-        time.sleep(1)
-        call(f'chmod 755 {cls.SCRIPTS_DIR}/scripts_{job_id}.sh', shell=True)
-        slurm_id = check_output(f'sbatch {cls.SCRIPTS_DIR}/scripts_{job_id}.sh', shell=True)
-        slurm_id = slurm_id.decode().strip().split()[-1]
-        logger.setLevel(logging_level)
-        logger.info(f'Queue: {job_name}_{job_id}, Job ID: {slurm_id}, Node: {node}, N_jobs: {len(cmds)}')
+        command = ['chmod', '755', script_path]
+        _ = run(command, stdout=PIPE, stderr=PIPE, universal_newlines=True)
+        command = ['sbatch', script_path]
+        result = run(command, stdout=PIPE, stderr=PIPE, universal_newlines=True)
+        slurm_id = result.stdout.strip().split()[-1]
+        logger.info(f'Sumbitted Task: {job_name} to {node}, containing {len(cmds)} jobs. Slurm ID: {slurm_id}')
 
-    def multi_submit(self, cmds: List[str], n_jobs: int, job_name: str, logging_level: int = logging.WARNING):
-        # TODO: support shuffling
-        cmd_bin = [[] for _ in range(min(n_jobs, self.pool_size))]
-        for i, cmd in enumerate(cmds, start=1):
-            cmd_bin[i % self.pool_size - 1].append(cmd)
+    def multi_submit(
+        self,
+        cmds: List[str],
+        job_name: str,
+        logging_level: int = logging.WARNING,
+        shuffle: bool = False,
+        sleep_time: float = 0.5,
+    ):
+        if shuffle:
+            import random
+
+            random.shuffle(cmds)
+        logger.setLevel(logging_level)
+        logger.info(f'Found {len(self.nodes)} available nodes.')
+        self.pool_size = min(self.pool_size, len(cmds))
+        logger.info(f'{len(cmds):,} jobs to excute, allocated to {self.pool_size} tasks.')
+        logger.info(f'Each task will use {self.ncpus_per_job} cpus.')
+
+        used_nodes = {}
+        registed_jobs = 0
+        for k, v in self.jobs_on_nodes.items():
+            used_nodes[k] = v
+            registed_jobs += v
+            if registed_jobs < self.pool_size:
+                continue
+            elif registed_jobs == self.pool_size:
+                break
+            else:
+                used_nodes[k] -= registed_jobs - self.pool_size
+                break
+        logger.info(f'Used {len(used_nodes)} nodes.')
+        logger.info(f'Each node will excute {self.max_jobs_per_node} tasks in parallel.')
+        logger.info(f'{used_nodes}')
+        k, m = divmod(len(cmds), self.pool_size)
         ith = 0
-        for node, n_jobs in self.jobs_on_nodes.items():
+        task_log = {}
+        for node, n_jobs in used_nodes.items():
+            logger.info(f'{node}: {n_jobs} tasks')
             for _ in range(n_jobs):
+                time.sleep(sleep_time)
+                start, end = ith * k + min(ith, m), (ith + 1) * k + min(ith + 1, m)
+                logger.info(f'Task {ith}: containing job {start}-{end-1}')
+                task_name = f'{job_name}_{ith:>03}'
                 self.single_submit(
                     self.nodes[node]['partition'],
                     node,
                     self.ncpus_per_job,
-                    cmd_bin[ith],
-                    job_name,
-                    f'{ith:>03}',
+                    cmds[start:end],
+                    task_name,
                     logging_level,
                 )
                 ith += 1
-                if ith >= len(cmd_bin):
-                    break
-            else:
-                continue
-            break
+                task_log[task_name] = {
+                    'node': node,
+                    'script': f'{task_name}.sh',
+                    'stdout': f'{task_name}.out.log',
+                    'stderr': f'{task_name}.err.log',
+                    'cmd': cmds[start:end],
+                }
+        with open(f'{self.FILE_DIR}/{self.time_now}.log', 'w') as f:
+            logger.info(f'Writing task log to {self.FILE_DIR}/{self.time_now}.log')
+            json.dump(task_log, f, indent=4)
 
     def starmap(self, func: Callable, params: Iterable[Iterable]):
         """
@@ -192,7 +244,7 @@ class SlurmPool:
         return: None
         """
         cmds = [func(*i) for i in params]
-        self.multi_submit(cmds, len(params), func.__name__)
+        self.multi_submit(cmds, func.__name__)
 
     def map(self, func: Callable, params: Iterable):
         """
@@ -201,7 +253,7 @@ class SlurmPool:
         return: None
         """
         cmds = [func(i) for i in params]
-        self.multi_submit(cmds, len(params), func.__name__)
+        self.multi_submit(cmds, func.__name__)
 
     @classmethod
     def clean(cls):