+++ CHANGELOG.md
@@ -55,7 +55,7 @@ Here we would have the update steps for 1.2.4 for people to follow.
 
 ### Fixed
 
-## [0.1.4] - 2022-12-17
+## [0.1.4] - 2022-12-18
 
 ### Added
 
@@ -65,5 +65,9 @@ Here we would have the update steps for 1.2.4 for people to follow.
 - Use jinja2 as sbatch template
 - change script dir to .autosbatch
 - get rid of pandas
+- replace call with subprocess.run
+- speed up by remove append
+- Add a logger
+- save job information to a file
 
 ### Fixed